requirements.txt
asgiref==3.2.3
boto3==1.12.9
botocore==1.15.9
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-appconf==1.0.3
django-autocomplete-light==3.5.0
django-ckeditor==5.8.0
django-filter==2.2.0
django-js-asset==1.2.2
django-pagination==1.0.7
django-ratelimit==2.0.0
django-staticfiles==1.2.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sql==0.4.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
